… ARMeilleure/Signal/UnixSignalHandlerRegistration.cs
… ARMeilleure/Translation/PTC/Ptc.cs
… ARMeilleure/Translation/Translator.cs
… Ryujinx.Audio/Renderer/Server/AudioRenderSystem.cs
… Ryujinx.Audio/Renderer/Server/AudioRendererManager.cs
… Ryujinx.Common/Configuration/MemoryManagerMode.cs
… Ryujinx.Cpu/CpuContext.cs
… Ryujinx.Cpu/IVirtualMemoryManagerTracked.cs
… Ryujinx.Cpu/MemoryEh.cs
… Ryujinx.Cpu/MemoryEhBase.cs
… Ryujinx.Cpu/MemoryEhMeilleure.cs
… Ryujinx.Cpu/MemoryEhUnix.cs
… Ryujinx.Cpu/MemoryEhWindows.cs
… Ryujinx.Cpu/MemoryManager.cs
… Ryujinx.Cpu/MemoryManagerBase.cs
… Ryujinx.Cpu/MemoryManagerHostMapped.cs
… Ryujinx.Cpu/Tracking/CpuMultiRegionHandle.cs
… Ryujinx.Cpu/Tracking/CpuRegionHandle.cs
… Ryujinx.Cpu/Tracking/CpuSmartMultiRegionHandle.cs
… Ryujinx.Graphics.Gpu/Engine/Compute.cs
… Ryujinx.Graphics.Gpu/Engine/MethodCopyBuffer.cs
… Ryujinx.Graphics.Gpu/Engine/MethodDraw.cs
… Ryujinx.Graphics.Gpu/Engine/MethodUniformBufferBind.cs
… Ryujinx.Graphics.Gpu/Engine/MethodUniformBufferUpdate.cs
… Ryujinx.Graphics.Gpu/Engine/Methods.cs
… Ryujinx.Graphics.Gpu/GpuContext.cs
… Ryujinx.Graphics.Gpu/Image/Pool.cs
… Ryujinx.Graphics.Gpu/Image/TextureGroup.cs
… Ryujinx.Graphics.Gpu/Memory/Buffer.cs
… Ryujinx.Graphics.Gpu/Memory/BufferCacheEntry.cs