This PR adds an additional type of memory manager, the HostMappedMemoryManager, which maps directly into a reserved "address space" range in the host memory, rather than into an emulated physical memory and accessing it with a software page table. Kernel memory management has been refactored to support this.

This PR is a joint effort - gdkchan did all the kernel refactoring, memory instruction changes and new memory manager, and I made changes necessary to make the memory tracking function performantly with host memory protection, and the Windows 10 emulated shared memory using placeholders.

More information on these changes and their implications will be posted in a comment later. Check the commit history for more detailed attribution.

For the record, I didn't come up with the name. 👼 

# Technical overview
## Current approach
One area of Ryujinx CPU emulation that is in dire need of improvements right now is memory access. Currently, guest memory access is implemented using inline software translation. This means that on each memory access, we need to do address validation, address translation and software protection for memory tracking. This approach has some benefits, but it also has some issues.

### Main benefits:

- Enhanced control over the memory mapping and address translation process. Allows things like guest shared memory to be easily implemented on all OSes.
- Memory reprotection is fast. This benefits GPU emulation as we need to track CPU writes to memory regions that the GPU can access, to invalidate data on GPU memory.

### Notable downsides:

- Poor performance, as it needs to produce roughly 20 x86 instructions per single Arm memory access instruction.
- Inflated code size, as all those instructions end up in memory (and also on disk if PPTC is enabled). This contributes to "JIT Cache Exhausted" crashes.
- Slow boot times because we throw more code into the pipeline, the JIT takes more time which means longer loads and more stutters without PPTC (first run).
- Slow PPTC build times (with PPTC), for the same reason mentioned above.

## How can we improve this?
We can take advantage of the host CPU MMU to make the address translation for us, instead of implementing it in software. This can be done by replicating the guest memory mappings on the host. Basically, we reserve a region of memory in the emulator process that can fit the entire guest address space, at a memory location "BASE", and then for each map that the guest tries to do at an address "A", we can just map "BASE + A" on the host. Similarly, on the CPU emulator, for each access to access "A", we just need to access "BASE + A".

The changes accomplish this by introducing a new memory manager called "HostMappedMemoryManager", which functions as mentioned above. The old "MemoryManager" is still there as there are some cases where it can be useful, such as situations that require memory aliasing to work (like services that need to access memory shared with other processes, if we ever decide to run services on the emulator).

As mentioned before, one of the advantages of the software translation is fast memory reprotection. With this new approach, we need to do host memory reprotection instead, which is considerably slower. For this reason, a few GPU emulation changes were required to avoid a large performance penalty associated with the reprotections. Constant buffer updates are now batched together and sent directly to the GPU, in addition to being written to CPU memory. Previously it would just write to CPU memory, which dirties the memory region and forces a reprotection, something we want to avoid now.

# Improvements
This change brings several improvements, which can be separated into 2 categories: performance and boot time. There are also a few other improvements that will be discussed later but, for direct comparison, these are the categories we will be focusing on.
The tests were performed on low and high-end hardware, on Windows and Linux, using all 3 GPU vendors (NVIDIA, AMD and Intel). All tests were performed on the most demanding locations on those games, to simulate a torture test/worst case scenario. As such, you can expect even better performance in other areas of the game.

## Performance
The comparison was done by calculating the average FPS on a single run of the game. Note the following configurations:

- **Ryzen 7 5800X** and **Intel i5-6600K** are equipped with NVIDIA GPUs.
- **Intel i5-1135G7** is using Intel Iris Xe graphics on Windows. (typically GPU bottlenecked)
- **Intel i7-4770** is using an RX580 on Linux (Pop OS).

### Locations tested:

![image](https://user-images.githubusercontent.com/6294155/118205301-edd13f80-b457-11eb-8caa-43038dd69cdb.png)
![image](https://user-images.githubusercontent.com/6294155/118205344-08a3b400-b458-11eb-832c-fdfad10aa046.png)
![image](https://user-images.githubusercontent.com/6294155/118205355-0fcac200-b458-11eb-85af-43a55fa7940c.png)
![image](https://user-images.githubusercontent.com/6294155/118205376-1fe2a180-b458-11eb-9682-431cc176d3cb.png)


- **Super Mario Odyssey:** Seaside Kingdom. Abnormally low FPS on this kingdom was a common complaint from users, so we made sure to get this one tested.
- **Animal Crossing New Horizons:** Heavily loaded island with a lot of buildings and NPCs, among other things. Kindly provided by user Leischii|Yannick. A simpler or empty island should have much better performance.
- **Super Smash Bros Ultimate:** Fountain of Dreams with 8 players, level 9. This has the highest load on the CPU. Matches with less players will have even better performance.
- **Splatoon 2:** Tested on Inkopolis Square. This has the highest number of NPCs and normally the lowest FPS anywhere in the game. Note that this specific area is locked to 30 FPS on the Switch; we had to uncap it for testing purposes as it was already hitting 30 with those changes on all PCs tested.
- **Mario Kart 8 Deluxe:** Baby Park track with CPU racers on. This track seems to get the lowest FPS compared to other tracks.

The oldest CPU of the bunch (a stock i7-4770 running Linux) had one of the highest improvements, averaging 50% FPS improvement across all the 5 games tested. (1.5x)

### Boot times
One common complaint is that the boot/load times are too slow. In fact, it’s so slow that some users assume that the game is not working and subsequently gives up, closing the emulator before the game finishes booting. These changes largely address this issue by significantly improving boot times, with both PPTC on and off. Below you can find some measurements.

#### PPTC Off ("first run/impression")
![image](https://user-images.githubusercontent.com/6294155/118205035-6d124380-b457-11eb-98ad-ce7cd24685f5.png)
![image](https://user-images.githubusercontent.com/6294155/118205048-73082480-b457-11eb-9592-eb70cf5b9372.png)
![image](https://user-images.githubusercontent.com/6294155/118205062-7ac7c900-b457-11eb-9a7b-778368776404.png)
![image](https://user-images.githubusercontent.com/6294155/118205075-81eed700-b457-11eb-8bbd-d88930e255cd.png)

#### PPTC On
![image](https://user-images.githubusercontent.com/6294155/118205150-a34fc300-b457-11eb-8ede-d1407c2597d1.png)
![image](https://user-images.githubusercontent.com/6294155/118205172-ad71c180-b457-11eb-8ece-cbf72023b89d.png)
![image](https://user-images.githubusercontent.com/6294155/118205132-99c65b00-b457-11eb-965a-57ffe1c4b145.png)
![image](https://user-images.githubusercontent.com/6294155/118205191-b4003900-b457-11eb-874e-64a2d715ebfe.png)

Not only is boot time greatly reduced but, with this performance update, some games may boot with PPTC disabled in the same amount of time as on master with PPTC enabled/cached. This means that some users may now wish to simply disable PPTC to avoid the occasional recompilation and other issues associated with it. For those that still wish to keep PPTC enabled, the performance is better than ever.

But that's not all. For PPTC users, having to recompile PPTC is a common annoyance. This update also improves that. Let's check out how long it takes to recompile PTC now:

![image](https://user-images.githubusercontent.com/6294155/118207177-02173b80-b45c-11eb-9b3e-e479f89c06cc.png)

As we can see, it takes less than half of the time to recompile the code on this game! On top of that, the PPTC cache size on disk is a lot smaller, consuming less than half the size on average. **Thanks to the lower size, the "JIT Cache exhausted" crash that has been plaguing some games should also be fixed.**

### First impressions are everything

Something that is harder to show on paper is the difference this update brings in those moments between menus and gameplay, when the game is first “spooling up” and may appear sluggish. This can manifest as low FPS or stutters unrelated to shaders, slow movement speed, and even (temporary) graphical glitches until the emulator catches up to full speed. Here's an example filmed by gdkchan of what this first impression looks like in Super Mario Odyssey, were you to boot a save from the Cascade Kingdom without having any PTC built:

**Before:** https://streamable.com/i7qcet
**After:** https://streamable.com/l9zrhk

### One more thing...
For Fire Emblem: Three Houses players, one long standing issue was the large character movement slowdown in the monastery. During our testing, we discovered that the issue was also fixed with this update. Check it out below.

**Before:** https://cdn.discordapp.com/attachments/787101806085406724/834130073417809970/2021-04-20_11-29-46.mp4
**After:** https://cdn.discordapp.com/attachments/787101806085406724/834131235952525372/2021-04-20_11-24-55.mp4

Similarly, other games with sluggish performance such as Hyrule Warriors: Age of Calamity have greatly improved. This game is known for skipping frames on the switch when it can't keep up, but Ryujinx's _cpu_ was too slow to handle the most intense of scenes in the game, so the game forcibly slowed time instead. 

**Before:** https://i.imgur.com/liFOjmz.mp4
**After:** https://i.imgur.com/Gkk9G6Z.mp4

# Configuring
![](https://cdn.discordapp.com/attachments/466000230987464719/841079111576715334/unknown.png)

- A new setting called "Memory Manager Mode" was introduced to allow selecting the memory management that should be used. "Software" is the current memory manager without any performance enhancements, while "Host" and "Host unchecked" are the new options added in this update. 
- "Host unchecked" has the best performance, however it doesn't boundaries-check the memory access, which makes it unsafe. It is recommended that you only use it with code that you trust. So, it should be fine if you are just playing games from your Switch or testing your own homebrew. 
- "Host" is a little bit slower, but ensures that the memory access is within the guest address space, which is safer. "Software" is still the most accurate one, but also the slowest. 

"Host unchecked" is the one enabled by default as it provides the best performance.

# Acknowledgements & what to test
All the tests and graphs presented here were made by EmulationFanatic, huge thanks to him for that!

We are looking for any regressions that this change might introduce on games, so if you find any, please let us know. If you do find an issue with the PR, be sure to test the game more than once on both master and this PR, to ensure that it is really a regression.

Any performance measurements are also welcome, so that we have an idea of the improvement here on a wider range of different hardware.

## Ac_Knowledgements
The graphs in this PR description somehow materialized from somewhere, and we're left forever wondering how they came to be in this world. Will we ever find out the origin of these graphs? Maybe you could find out for yourself, if you know who to Ac_K.

# References
All tests performed with Ryujinx default settings except where otherwise noted (e.g. vsync off or pptc off).

# Full specifications of PCs tested:

**Ryzen 7 5800X @ 4.8GHz**
32GB DDR4 3600
NVIDIA GTX 1080
Windows 10 20H2

**Intel i5-6600K @4.1GHz**
16GB DDR4 2400
NVIDIA GTX 1080
Windows 10 20H2

**Intel i7-4770 (stock clocks)**
16GB DDR3 1600
AMD RX580
Linux - Pop OS 20.10 with TKG PDS kernel (latest stable)

**Lenovo T14 laptop
Intel i5-1135G7**
16GB DDR4 3200
Intel Iris Xe graphics
Windows 10 20H2