@AcK77 2021-05-14
Download the artifacts for this pull request:

* [ryujinx-Release-1.0.0+befc23a-linux_x64](https://nightly.link/Ryujinx/Ryujinx/actions/artifacts/60513085.zip)
* [ryujinx-Release-1.0.0+befc23a-osx_x64](https://nightly.link/Ryujinx/Ryujinx/actions/artifacts/60513086.zip)
* [ryujinx-Release-1.0.0+befc23a-win_x64](https://nightly.link/Ryujinx/Ryujinx/actions/artifacts/60513087.zip)