@Shahil-Ayato 2021-05-14
One issue is using the 6gb Dram hack will cause the two new memory manager modes to crash