@Illynir 2021-05-14
PC:

Ryzen 2700 @ 4Ghz
16 GB  DDR 3000Mhz
Nvidia RTX 2060
Windows 10 20H2

Game tested:

I took the most " demanding " areas of each game, without upscaling or anisotropic.
Master => PR with Host unchecked

- Xenoblade 1: 23 FPS => 21 FPS - No more Slowdown on the character if you are not fullspeed.
- Xenoblade 2: 10-25 FPS => 30 FPS - FPS no longer fluctuate, no more slowdown on the character if you are not fullspeed.
- FE3H : 10-15 FPS => 30 FPS - No more slowdown on the character if you are not fullspeed.
- Hyrule Warriors: Age of Calamity : 10 FPs => 14 FPS - Sound is better, no more slowdown on the character if you are not fullspeed.

Beyond the performance, the good thing is that the loadings are much faster than before and that the fluctuations of the FPS are much less. XC1 has lost 2 FPS but the game seems more "fluid" nonetheless.