@Illynir 2021-05-14
- **Buried stars** : 52 FPS => 60 FPS - It's an interactive visual novel so it was already at 60 FPS except at a specific place in the menu before playing, I did my test here.
- **Monster hunter Rise** : I didn't see any difference but it was on master and PR, without Macro HLE and without the optimizations. I guess it makes a bottleneck.

I can't test the performance gains on these games because they were already fullspeed for me but there is no regression found on:

- **Astral Chain**
- **Bravely Default 2**
- **Famicom Detective Club (Both)**