@zzgundamx 2021-05-15
Encountering this crash on the latest PR. However PR 07fc3dd works fine. I've also tried extracting to a new folder and copying over my Portable directory too.

[Ryujinx_1.0.0+fa6231b_2021-05-15_08-36-51.log](https://github.com/Ryujinx/Ryujinx/files/6481644/Ryujinx_1.0.0%2Bfa6231b_2021-05-15_08-36-51.log)