@zzgundamx 2021-05-15
[Ryujinx_1.0.0+07fc3dd_2021-05-15_09-08-00.log](https://github.com/Ryujinx/Ryujinx/files/6481689/Ryujinx_1.0.0%2B07fc3dd_2021-05-15_09-08-00.log)

This is the log that I have for 07fc3dd - if it helps. There are no errors on this build.