@EmulationFanatic 2021-05-15
> Encountering this crash on the latest PR. However PR [07fc3dd](https://github.com/Ryujinx/Ryujinx/commit/07fc3dd26e8f47d44194cfd5abe67f824afdf251) works fine. I've also tried extracting to a new folder and copying over my Portable directory too.
> 
> [Ryujinx_1.0.0+fa6231b_2021-05-15_08-36-51.log](https://github.com/Ryujinx/Ryujinx/files/6481644/Ryujinx_1.0.0%2Bfa6231b_2021-05-15_08-36-51.log)

@zzgundamx Can you re-enable all standard logs and upload the results again on both the latest commit and the one before?