@Powerlated 2021-05-15
Should probably change (fast, unsafe) to (fastest, unsafe) to be a little bit more clear.