@gdkchan 2021-05-15
> 
> 
> Should probably change (fast, unsafe) to (fastest, unsafe) in the settings window to be a little bit more clear.

It's already "fastest", the image on the PR description is outdated.