@riperiperi 2021-05-16
I've made a new change that makes UBO updates batch together a lot better, without having to reprotect every time like master does. **Make sure to test games again in case anything breaks, gets faster or gets slower.**

**Known improved titles:**

- Breath of the Wild (about _1.5x_ improvement from older PR version in average case. best case improves significantly over master)
- Animal Crossing (improved worst case when FIFO% is high)
- Super Mario Odyssey (improved performance in Seaside Kingdom, Lost Kingdom)
- Other FIFO limited games may experience a small FPS improvement