@zzgundamx 2021-05-16
> One issue is using the 6gb Dram hack will cause the two new memory manager modes to crash

I've tried using the 6gb Dram hack and so far it has been fine with the memory manager set to Host (safe). Albeit I've been only playing one game so far (gdkchan's fd7a459 build).