@AcK77 2021-05-16
6GB DRAM mode issue is fixed since 7207898