@SandeMC 2021-05-17
Played a little on my 2200G which is pretty far from being a good CPU and 8GB RAM which is also far from being good, the only game I have is Super Mario Odyssey,  so, Cap Kingdom is mostly 60 fps with exception of spawn point till bridge being 45 for some reason, Cascade Kingdom mostly 60 fps with dropdowns to 50-55 when waterfall is rendered, Sand Kingdom mostly 60 fps with the exception of when Tostarena is rendered, then it's 40-45 fps. Didn't test other kingdoms(sorry, I'm lazy)