@brunbrunbrun 2021-05-17
Using a i7-10700k with 32gb of ram and a 1660S, tested a few games on  [d7f4df7](https://github.com/Ryujinx/Ryujinx/commit/d7f4df7aed36922cde280b365b75ee51facefe4d) with **Host unchecked**:

- **Animal Crossing: New Horizons** - Trouble with first time boot, crashed after a few moments, after 3 tries it worked normally.
- **Arms** - Heavy stutter at the start of any match, but after that it stays at 60fps constantly; It would previously have no said stutter on the start, but would fluctuate in the 50~55fps ( a worthy tradeoff imo).
- **Astral Chain** - Now runs at a stable 28~30fps; Before, it would be on the low 20s.
- **Fire Emblem: Three Houses** - Load times got slightly faster, MC movement around the monastery felt more fluid, and framerate was more constant; One problem tho, after a few loading screens, HUD elements (especially textboxes) would start flickering offset, mildly infuriating, but fixed with a restart; Happened most times when there were a couple changes in maps (as in monastery -> battle -> monastery).
- **Monster Hunter Rise** - No noticeable differences on my end. 
- **Pokémon Sword** - More stable framerate in less demanding areas; And the weird geometry lines being drawn over textures, was not present anymore. Before:
![firefox_XjzDSP3YFC](https://user-images.githubusercontent.com/82368316/118566124-6a5a6b80-b741-11eb-8374-ab8542b81488.png)

- **The Legend of Zelda: Breath of the Wild** - No noticeable differences on my end.

Overall, it was a huge improvement all around.
Nice work.