@Benedani 2021-05-18
Does using host checked instead fix these issues? If so, I feel like it should be the default instead of host unchecked, since it caused a crash for me while loading into Super Mario Odyssey. The crash had to do with OpenAL??