@Linear524 2021-05-18
Testing with Xeon 2690V4 @ 32GB Ram and 1080 with **Host Unchecked mode**:

**Luigis Mansion 3** = from Master 20~27 --> to stable 30fps with this PR  (reflection bug is still here but not related to this PR)
**Astral Chain** = Master 12fps  --> PR 27~30fps
**Tourist** = Master 20fps  --> 23fps  (FIFO bottleneck)
**Zelda BoTW** = Master 20fps  --> PR 20fps  (stuck at 20 fps by strange game logic)
**SMO** = Master 25~30fps  --> PR 30fps always !
**Spyro Reignited** = Master 22~25fps  --> PR Stable 30fps !!

No additional issues so far with this PR, maybe port it to Master ?