@riperiperi 2021-05-18
> 
> 
> Does using host checked instead fix these issues? If so, I feel like it should be the default instead of host unchecked, since it caused a crash for me while loading into Super Mario Odyssey. The crash had to do with OpenAL??
> *didn't verify if it still happens on current commit

Host (checked) does not have fewer crashes than Host Unchecked. What it guards against is unsolicited modification of Ryujinx program memory outside of the game's emulated memory, which can only happen if the game is doing an out of bounds memory access anyways (will crash 99% of the time if we check and we don't). The concern is programs that are aware of this and attempt to do something malicious, like homebrew programs specifically crafted as an attack.